# Use fast-float to parse floats for 50-75% speedup