Came across
https://lemire.me/blog/2021/01/29/number-parsing-at-a-gigabyte-per-second/
which pointed out the rust library. Since we recently optimized this because it
was showing up in a customer profile I thought it'd be worth doing:

    $ cargo bench -- parse_float
    parse_float32/-3.0      time:   [10.227 ns 10.292 ns 10.364 ns]
                            change: [-74.777% -74.335% -73.827%] (p = 0.00 < 0.05)
                            Performance has improved.
    Found 8 outliers among 100 measurements (8.00%)
      1 (1.00%) low mild
      3 (3.00%) high mild
      4 (4.00%) high severe

    parse_float32/9.7       time:   [9.7706 ns 9.8281 ns 9.8885 ns]
                            change: [-76.467% -75.562% -74.784%] (p = 0.00 < 0.05)
                            Performance has improved.
    Found 7 outliers among 100 measurements (7.00%)
      2 (2.00%) low mild
      3 (3.00%) high mild
      2 (2.00%) high severe

    parse_float32/NaN       time:   [6.3433 ns 6.3743 ns 6.4075 ns]
                            change: [-69.361% -69.100% -68.846%] (p = 0.00 < 0.05)
                            Performance has improved.
    Found 10 outliers among 100 measurements (10.00%)
      7 (7.00%) high mild
      3 (3.00%) high severe

    parse_float32/inFiNiTy  time:   [9.1302 ns 9.1807 ns 9.2367 ns]
                            change: [-50.457% -50.038% -49.601%] (p = 0.00 < 0.05)
                            Performance has improved.
    Found 10 outliers among 100 measurements (10.00%)
      6 (6.00%) high mild
      4 (4.00%) high severe

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/materializeinc/materialize/5517)
<!-- Reviewable:end -->